+++ CHANGELOG.md
@@ -23,6 +23,7 @@ omnibus-gitlab repository.
 - Update Mattermost to 5.4.0
 - Warn users when postgresql needs to be restarted due to version changes !2823
 - Ensure postgres doesn't automatically restart when it's run file changes !2830
+- Add nginx-module-vts for Prometheus metrics collection !2795
 
 11.4.1